+++ RELEASES.md
@@ -8,11 +8,21 @@ as-is, without any warranty. -->
 Version 1.7.0 (unreleased)
 ==========================
 
+  * The following methods were added to all fixed-point numbers:
+      * [`from_be`][f-fb-1-7], [`from_le`][f-fl-1-7]
+      * [`to_be`][f-tb-1-7], [`to_le`][f-tl-1-7]
+      * [`swap_bytes`][f-sb-1-7], [`reverse_bits`][f-rb-1-7]
   * For the experimental feature [`num-traits`][feat-exp-1-7], the
     following traits were implemented where applicable:
       * [`OverflowingAdd`][nt-0-2-oa], [`OverflowingSub`][nt-0-2-os],
         [`OverflowingMul`][nt-0-2-om]
 
+[f-fb-1-7]: https://docs.rs/fixed/~1.6/fixed/struct.FixedI32.html#method.from_be
+[f-fl-1-7]: https://docs.rs/fixed/~1.6/fixed/struct.FixedI32.html#method.from_le
+[f-rb-1-7]: https://docs.rs/fixed/~1.6/fixed/struct.FixedI32.html#method.reverse_bits
+[f-sb-1-7]: https://docs.rs/fixed/~1.6/fixed/struct.FixedI32.html#method.swap_bytes
+[f-tb-1-7]: https://docs.rs/fixed/~1.6/fixed/struct.FixedI32.html#method.to_be
+[f-tl-1-7]: https://docs.rs/fixed/~1.6/fixed/struct.FixedI32.html#method.to_le
 [feat-exp-1-7]: https://tspiteri.gitlab.io/fixed/dev/fixed/#experimental-optional-features
 [nt-0-2-oa]: https://docs.rs/num-traits/^0.2/num_traits/ops/overflowing/trait.OverflowingAdd.html
 [nt-0-2-om]: https://docs.rs/num-traits/^0.2/num_traits/ops/overflowing/trait.OverflowingMul.html